Add basic list command and usage
The list command just prints out the ID of all of the trading pairs and the base and quote currencies for each pair. It should be relatively easy to add moreo details. I want the subcommand list to be used to find the ID of a pair. That ID would then be used for the other subcommands (e.g. an alert command that sets custom price alerts).